refactor view-letter page
